+++ front-end/src/index.js
@@ -24,9 +24,9 @@ const theme = createMuiTheme({
       contrastText: '#000',
     },
     secondary: {
-      light: '#757ce8',
-      main: '#3f50b5',
-      dark: '#002884',
+      light: '#e57373',
+      main: '#ff7043',
+      dark: '#ff5030',
       contrastText: '#fff',
     },
   }